Grab content after executing logic
This does introduce another bug where a 404 error is generated in cases
where there is a view path but no logic. More thought is needed on how
to handle this, but my assumption is the 404 determination needs to be
made after Application and Presenter are called. Possibly in it's own
middleware.