Fix 404 logic
Properly handles all cases of routing/presenting taking into account
the all_views_visible config option. This requires a change to
presenter to allow resetting. Otherwise the content in the matching
view path would always be returned in the response, regardless of the
all_views_visible option.

This code also properly handles the case where a view is set in the 404
handler.