Always act on ViewComposer
Presenter#view now always returns the ViewComposer (unless the view has
been overridden in which case it returns a View). This forces the
developer to always interact with the parts rather than the whole.

This is transparent to the developer. ViewComposer allows the parts to
be treated as a whole, while waiting until the last possible moment to
actually compose the view.