Soft copy doc when matching
There was a bug when more data than views, where the reference to the underlying structure was being lost due to duplication. This led to the creation of `soft_copy`, where the structure remains the same while the node is duped. Useful in cases where you want a new node but it'll be part of the same underlying structure (like matching).