Search all view stores for matching view
Prior to this, only the current "working" store was queried. This change makes it possible for stores to define views that can later be overridden in other stores.

A use-case would be that some library registers a view-store with an app that contains a default view, which the app can override by creating a view at the same path.