Do not precompose views
This is an artifact from how views used to be built. By precomposing we allowed Pakyow to be more performant in cases where the view was unaltered prior to being rendered. With StringDoc, this is less important because it's fast -- so treating this as we would a premature optimization. And it also didn't quite work as-is.