Add post-processing to route template expansions
Adds a post-processing step to route template expansion to give the template an opportunity to make any last-minute changes to the routes. In the restful route case, we use post-processing to move the `show` route to the end of the defined list of routes to avoid conflicts with `new` and `collection` routes.

Refactored some of the route eval code in the process of coming up with a reasonable solution for this.