Include partials before matching
This resolves a wild edge-case around matching a scope containing a
partial that’s also defined in a partial. Previously, the props within
the scope were not visible because the view saw a comment, not a
partial to be included. Really this needs a better solution (documented
in the FIXME), but that would lead to a pretty significant effort.