Fix ui mutation bug
There were several issues here, the start of which is that mutations
weren’t unregistered when sockets disconnected. This caused multiple
mutations to be executed when a mutation occurred.

This resolves the issue by unregistering mutations when sockets
disconnect, avoiding the redundancy. It also pushing mutation messages
to specific clients rather than through a channel.