Move app methods to modules
Pakyow::App has been notoriously hard to test and overly complex from
the beginning. This commit is the first step towards improving this
part of the codebase.

The decision to move methods out of Pakyow::App into their own modules
was driven out of a desire to make the code a) more testable and b)
easier to understand. Each concern of the app is now contained in its
own module (e.g. configuring).

Everything a developer needs to know is available right there in the
module. Because there are no dependencies, the methods around a
particular concern can be tested in true isolation.