Refactor resource implementation
This change fits resources into the `Configuring` helpers to be
consistent with the other app config methods. It also adds an
instance-level `resource` method so that resources can be defined at
the class or instance level (just like routes). The tests are also
reorganized into unit and integration tests.

The only functionality change is `resource` no longer returns a route
set when no block is passed. That behavior was inconsistent, as
`resource` blindly calls each `RESOURCE_ACTION` and has no underlying
knowledge of the side-effects.