Fix edge case around StringDoc removal
We were resetting `@node` but that doesn’t change the “global” instance
of the node that might be present in other StringDoc objects. Now we
simply reset the state of the `@node` to be empty.

The other side of this is that we want to perform future modifications
to a removed StringDoc on the node, not the structure (since the node
is shared between instances and the structure is unique to this
particular StringDoc).