Refactor View#bind_attributes_to_doc
A couple of questions remain. It seems that we can refactor this further
in light of the addition of binding parts (since hashes won't be used).
Namely, values will always be procs, so we can remove some conditionals.
Also, it was necessary to add a line to get the real value from an
attribute value proc being called. In this way either the return value
or the value of the Attribute instance is used.