Remove type check when appending view to collection
This was causing things not to work right in some cases when using
pakyow-test, because the object being appended to the collection was an
`ObservableView` rather than a ` View`.

I considered adding the ability to register additional object types
that are valid in a collection, but need to think more about that.
There are currently a few places that a view can be added to a
collection, and this was the only one where we performed the check. So,
it was inconsistent.