Fix redis subscription bug
When the subscription is unsubscribed, it disconnects from Redis. This
keeps things nice and clean. The problem is that we weren’t creating a
new subscriber, only calling subscribe on a “dead” subscriber.

This change avoids this problem by effectively discarding the dead
subscriber and creating a new one every time we resubscribe to channels.