Always load logger middleware
When the middleware isn’t loaded, the request logger isn’t used. This
creates a problem because the logging interface is different (for
example, when the request fails a `houston` method is called on the
request logger). We now always load the middleware and the request
logger simply logs to /dev/null when logging is disabled.