Simplify app running
Pakyow apps now run only with a single environment. This was originally
implemented as a “we could run an app in multiple environments!” but
was never once used in the real world and wasn’t worth the complexity
that it brought.

Similarly was the care that we took to prevent an app from being
prepared, staged, or run more than once. In normal use this would never
ever happen, so once again it wasn’t worth the complexity.