Initial Pakyow environment implementation
Very much still a work in progress; tests are completely hosed at the
moment but I want to see this through to make sure the idea works. The
tests needed serious work anyway :/

The idea here is for Pakyow::App to not know how to run itself.
Instead, the Pakyow environment (contained within the Pakyow module) is
responsible for configuring environment-level things and running Pakyow
applications within it. This is the first of many steps towards making
it possible to have multiple Pakyow apps mounted within a single
environment.