Improve hierarchical error handling
With the previous approach, @current_router was only ever a top-level
router. This caused problems when the router to trigger the error was
not a top-level router. Now @current_router is set by the router itself
when routing, so we know exactly where we are in the routing process.

Also, @current_router is now always an instance.