Initial cut at presenter error presentation
It looks like we can leverage global handlers to provide some default
behavior for users. In development, we can handle MissingView and
provide insight to the user on what to do next. We can also provide
insight into any 500/404 errors that occur, just like we have in
previous versions of the framework.

We can also present built-in error views in development and fall back
to the app’s error views in production. This is possible by registering
Presenter’s error store only in development, but leaving the default
error handlers in place.

Once we introduce the concept of renderers, we can actually populate
the views with data. Until then, error views will appear to be broken.