Use binders in presenter
When data is presented by a presenter, it is decorated with the binder
for the scope. Binders being a concern of presenter rather than view
will allow us to further simplify, since views no longer have to
concern themselves with context, or even that a binder exists; it’s
just data now.