Initial reimplementation of view versioning
Presenting across a set of views is something we will not support for v1.0. It is not something we have used in practice, while it adds a lot of complexity both in implementation and in usage. Previous implementations have not been implemented correctly and the time it would take to do so is better spent on core features.

In light of this, VersionedView replaces ViewSet, which means that multiple nodes of the same significance that are "found" together are treated as versions of the same view. The `find_all` method will later be implemented to return an array of found views rather than the single View or VersionedView that `find` returns.