Fix bug when inheriting configured objects
We cache off the parent, or the configurable object, when creating a config group. This was creating a problem when the configurable object is subclassed, and the config value is a proc that relies on object-level context. It would be incorrectly called in context of the original configurable object instead of the subclass.

This commit resolves the problem by setting the new parent.