Refactor core into a framework
This commit separates routing concerns into a separate framework. Pakyow::App now has the basic knowledge it needs to respond to requests; everything else is contained in registered endpoints. For example, the routing framework registers a Controller endpoint that is called with details of the request.

It's now possible to load an app without the routing framework, replacing it with something else. This wasn't possible before because we extended core classes. Now, subclasses for each object (e.g. app, controller) are created along with the app and behavior is loaded into those subclasses only. This means that two apps running in the same process have their own namespaces, with different behavior.

A related change is that the controller/router concepts have been combined into a single controller concept. Prior to this commit, controller was an internal pakyow object and router was the thing defined by the end user that contained backend logic. That router has become the controller, which is more clear. The internal controller object was refactored away, simplifying things overall.