Replace controller hooks with pipelines
The current hook implementation made it quite difficult to install default behavior in the controller class and let subclasses override, replace, or skip that behavior. While considering the options, I decided to simplify the entire thing. Pakyow has always internally represented a route as a pipeline of functions, so this was a pretty natural change.

We lose the ability to define actions to occur after a route. This is fine, because I don't think I've ever had a use for after hooks. They also had the appearance of being unreliable, because they are never called after a halt. Pipelines now always call the route last, unless an action has halted the request.

Anyway, this sets us up to move csrf protection into the controller and let controllers disable that behavior if they would like (e.g. for an api).