Refactor routing to rely more on pipelines
There's a lot going on here so I'll try to summarize.

Each controller is now registered as a pipeline action. When a controller receives a request, it looks for a matching route. If it finds one, it dups (so we can safely set instance state) and then calls the route. Calling is handled by the pipeline.

Routes now do not keep up with their own pipeline. Instead, all possible actions are part of the default controller pipeline. Before calling, the controller modifies the pipeline for the route being called, removing actions as necessary. The route itself is called through a `dispatch` action, which allows us to call the route block as part of the controller's pipeline without defining a route method.

This work provides better separation of concerns, which was the overall goal. Controller now handles errors that occur in a controller, with the app handling errors that occur outside of the controller. Missing is just a fall-through state now, with the app again letting controller handle those cases.

Rendering is better separated as well. Implicit rendering occurs as a finalize hook on the connection object, but only when the route goes through a controller. In prototype mode (when the controller is not loaded), the request is handled by an AutoRender action. It works out beautifully.