Let environment deep freeze on boot
It's hard to overemphasize how much clarity this brings. Before, Definable froze on definition and it created a lot of indirection. When trying to make Configurable freeze the config after configuring, things became even more complicated so I stepped back and this is the result. The driving idea here is that we don't want state to change after boot. Before then it's fair game.