Proxy requests in development server
Prior to this commit, requests made when the server is restarting would fail.

This work resolves the issue by starting a persistent server on the environment port that forwards requests to the current environment server process. When a restart occurs, the environment server process is replaced but the proxy server sticks around to receive the request. The proxy waits 15 seconds for the environment server process to exist, then times out.