Introduce rescue mode for apps
When an app fails to start, it enters rescue mode. This causes the app to log and respond with the initialization error. I added this primarily so that development is nicer. If you make a mistake, you can fix it, save, and the server will automatically restart itself. Without this feature, you'd have to go back to the terminal to start the server yourself.