Remove form-related code from presenter, create subclasses
Form code is being moved into its own framework / gem. Much of the functionality I want to bake in (e.g. error handling) relies on the ui and data frameworks. Making presenter dependent on those two frameworks for a feature you may not choose to use in an app seems like a mistake.

Something that fell out with this is moving ViewPresenter functionality into Presenter. We don't gain anything by keeping them separate and it makes presenter harder to explain.