Initial ui specs
This introduces a pattern that will let us confirm that applying transformations on the client leads to the same result as doing a full re-render on the server.

Each ui case is serialized in a folder under pakyow-js that contains the initial render, the transformation, and what the result of the transformation should look like.

In pakyow.js, we can automatically test these cases by applying the tranformation to the initial render and comparing with the result. If they don't match, there's a bug on the ui side.