Remove resubscribe pattern from data subscriptions
WIP, but I think with a change to how we handle data updates we can simplify data subscriptions quite a bit, fixing some other bugs that I found in the meantime.

We create a lot of work for ourselves by keeping up with the underlying object ids. Now the sole reason for this (based on the thinking captured in specs) is to catch cases where an object is changed in a way that causes it to no longer be covered by the query. When that happens the qualifications don't match, so it's missed unless we see that it's an object we're subscribed to.

What I plan to attempt is to change update so that it fetches the objects to update, then executes the query, then updates the local objects in place with the new values, keeping up with what the old values were as well as the values that were changed. That gives us all the information we need to check if the object was covered, or is now covered--either of which we want to trigger an update for.