Refactor ui presenter; support endpoints
Turned into a big change, because we need to track calls to internal presenter methods for supporting endpoints. Fortunately this did fix a slew of previously unknown bugs in the process, and also simplified pakyow.js quite a bit.

There's at least one tradeoff, which is socket traffic is more verbose due to `transform` and `bind` being tracked with the same arguments. There's likely an optimization we can do here in the future.