Refactor view presenter to be pipelined
This lets us register actions for a presenter, which can then be called in sequence. It lets us define the default presentation logic in a `perform` action that can be overridden by any subclass, while still allowing the default logic to be called through `super`.

To support, UI had to change. Rather than a UI Presenter that wraps the main presenter, we now create subclasses of each presenter on boot, then extend the ui-specific subclass with recording behavior. This lets the ui presenter operate just like the original, but everything will be recorded for serialization.