External scripts
This feature allows Pakyow developers to include external scripts without using a package manager like Bower or Yarn.

It is *not* a dependency manager. Instead, it avoids dependency management by downloads=ing the umd distribution version of each external script. Right now it uses unpkg.com as the provider, but this will be replaced with our own service at some point.

On boot, any external scripts that don't exist locally will be downloaded. The expectation is that they'll be checked into version control, so this will not run during a deploy (in fact fetch is completely disabled in production mode).

Pakyow.js is managed through this feature. When a new version is released, the user can simply run `pakyow assets:update` to get it.