Improve boot process
Solves a bunch of oddities when running in a forking web server like puma. Allows the main process to boot, then each worker to receive the `booted` callback and issue reconnects, etc. Still work to be done in reestablishing connections and such, but this commit sets us up for that.