Refactor environment boot and app finding
This commit makes it possible to boot the environment without running it. Used in cases like `pakyow irb` where apps need to be found and used, but won't be running in a server environment. Also changes `Pakyow.app` to find apps rather than define them when the environment has been booted.