Smarter realtime ui state timeouts
Allows sockets to be disconnected for a configurable time (default 24 hours) without losing state. Does not affect performance, since only "active" subscriptions (subscriptions with an active connection; meaning a connection non-expiring state) are processed. 24hrs of subscription data will be stored in Redis, rather than 60s, but this seems like the right tradeoff 90% of the time.