Add the navigable component to Pakyow.js
The Navigable component performs navigations in the background, replacing the body/head instead of a complete re-render.

Forms are also submitted in the background (if a websocket connection is active). If the submission successful, a navigation is triggered. If unsuccessful, the client does nothing and instead expects the server to push the transformations necessary to place the form in an errored state.