Wrap results returned from data sources
Results returned from data sources through the data proxy are now wrapped in a `Pakyow::Data::Result` object. This allows us to always track down how a particular result was fetched, even if a `nil` result was returned. The only time a return value is not wrapped is when a method ending in `?` is called. We return the raw boolean in this case so falseyness is not broken. On a related note, falseyness *is* broken for the `nil` result case, so `nil?` must be used to explicitly check for a `nil` result.

UI leverages this to handle transformations for previously unsupported `nil` values as well as counts.