Expose all values to components
This has proven necessary with forms. If the end user exposes an object to a form from a controller, running the app with the form framework will cause the form to no longer be setup for the object. We don't want a behind-the-scenes component to change the behavior for a user, so making this more permissive avoids the issue.