Handle stale uis
Pakyow.js receives the current app version from the websocket on connect. It tracks the version over time and when it changes, broadcasts on the `pw:ui:stale` channel. The `freshener` component listens on this channel and transitions itself to stale if the app version changes.

This works because app restarts will always cause websocket connections to reconnect.