Solve a fairly nuanced bug in presenter
The new "presentation_logic" test illustrates the problem that was occuring. This change simplifies versioned views so that transformations are always applied to every version contained within. It does this by initializing a versioned view with a single view instance that wraps a string doc meta node. It defers to meta node for applying changes across all versions.