Fix bug causing data subscriptions to not expire for a web socket
Using thread pools larger than 1 thread can cause changes to be applied out of order from time to time. While the executor guarantees that tasks are executed in order, it doesn't guarantee that one finished before another. This was causing odd and unpredictable behavior that was resolved by dropping down to a single thread executor.

The change to the after render hook guarantees reduces calls to the data adapter when multiple renders happen for a single request (e.g. component renders). This is safe since all renders are for the same socket id.

Unfortunately this is extremely difficult to verify reliably with a test. Fortunately this code doesn't have reason to change in the near future.