Run CI on external servers
GitHub workflows are failing intermittently due to lack of resources when spinning up MySQL, Postgres, and Redis containers. Something must have changed, because this started all of a sudden after running fine for some time.

This commit changes our workflows to run CI on external servers. For now we'll use inexpensive preemtible instances on Google Cloud Platform, but this gives us the flexibility to easily make a change in the future--either to a different provider or just resizing instances. I also expect this to be useful for automated performance testing.