Move version.rb into pakyow/core
version.rb is required in pakyow/core, but ships in the pakyow meta gem. This creates a dependency from pakyow/core to the meta gem, which has a dependency for every framework. Since pakyow/core is a dependency of every framework, it's impossible to include specific frameworks in a project without making all of them a dependency.

Moving version.rb to pakyow/core fixes this.