Bootstrap bundler without exec
The `pakyow` command automatically bootstraps bundler, preventing the user from having to add `bundle exec` to every command. Our previous approach was to replace the process with one that ran the binstub using `exec`. This ended up not being ideal for a couple of reasons:

1. It only worked if you ran `bundle install --binstubs`.
2. Replacing processes is slow.

This commit changes the approach by setuping up Bundler in the current process if a `Gemfile` is present. In my testing it saves ~200ms.