Setup default logger, global logger in the environment
It's sometimes necessary to log messages prior to the environment being setup. The environment now builds a default logger / global logger on demand, which is then replaced with a properly configured logger during setup. This ensures there's always a usable logger available no matter what state the environment is in.