Replace the thread local's internal logger instead of overwriting
This allows objects that hold a reference to the thread local logger before setup to still point to the right object and log to the appropriate logger after setup. A good example to think through is the environment deprecator. If the deprecator is used before setup, it is initialized with a reference to `Pakyow.logger`. When we overwrite the logger, the reference that the deprecator holds is the old logger instead of the new one. This change avoids the whole issue by replacing the thread local's internal logger instead of replacing the entire instance.