Deprecate Pakyow::Logger#silence
`Pakyow::Logger::ThreadLocal` supports silencing and is now the preferred path. It's safer because it avoids mutating state that could be shared across threads (the only mutated state is the values of the thread local). `Pakyow::Logger#silence`, on the other hand, mutates the logger instance itself and is not thread-safe. We might could make `Pakyow::Logger#silence` thread safe, but since the thread local logger is available everywhere it's much easier to support it there.